Add static_assert() pass to applicable tests

+++ test/test.h
@@ -5,9 +5,9 @@
 #include <stdlib.h>
 #include <assert.h>
 
-#define ASSERT(x, y) test_assert(x, y, #y)
+#define ASSERT(x, y) test_assert((int)(x), (int)(y), #y)
 #define DASSERT(x) static_assert(x); ASSERT(1, x)
-#define EASSERT(x,y) static_assert((x) == (y)); ASSERT(x, y)
+#define EASSERT(x,y) static_assert((int)(x) == (int)(y)); ASSERT(x, y)
 
 extern
 #if defined __cplusplus
